+++ crates/atlassian-swc-compiled-css/tests/fixtures/css-complex-selectors/out.js
@@ -1,0 +1,8 @@
+const _ = '._syaz5scu>span:first-type-of{color:red}';
+const _2 = '._syaz13q2>span:first-type-of:hover{color:blue}';
+import {ax, CC, CS} from '@compiled/react/runtime';
+const styles = null;
+<CC>
+  <CS>{[_, _2]}</CS>
+  {<div className={ax(['_syaz5scu _syaz13q2'])} />}
+</CC>;