+++ crates/atlassian-swc-compiled-css/tests/fixtures/css-imported-values/extract.js
@@ -1,0 +1,13 @@
+const _ = '._1bsb1yd6{width:var(--13gm3b0)}';
+const _2 = '._4t3i168g{height:var(--1v470gf)}';
+import {ax, ix} from '@compiled/react/runtime';
+import {value} from 'location';
+export const View = () => (
+  <div
+    style={{
+      '--13gm3b0': ix(value),
+      '--1v470gf': ix(value + 'px'),
+    }}
+    className={ax(['_1bsb1yd6 _4t3i168g'])}
+  />
+);