+++ packages/configs/webextension/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @atlaspack/config-webextension
 
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/core@2.26.2
+  - @atlaspack/config-default@7.0.2
+  - @atlaspack/packager-webextension@2.14.37
+  - @atlaspack/runtime-webextension@2.14.37
+  - @atlaspack/transformer-webextension@2.14.37
+  - @atlaspack/types@2.15.27
+  - @atlaspack/transformer-raw@2.14.37
+
+## 7.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`9cd9521`](https://github.com/atlassian-labs/atlaspack/commit/9cd9521978f783046e2ae4ce78f2de7aeb07d073), [`d0c7bea`](https://github.com/atlassian-labs/atlaspack/commit/d0c7bea04458255b4c5d4299600e734b5f919fe1), [`1086c6a`](https://github.com/atlassian-labs/atlaspack/commit/1086c6a2c59271b63b3fb78e6acdb9d3a0dbf354)]:
+  - @atlaspack/core@2.26.1
+  - @atlaspack/config-default@7.0.1
+  - @atlaspack/packager-webextension@2.14.36
+  - @atlaspack/runtime-webextension@2.14.36
+  - @atlaspack/transformer-webextension@2.14.36
+  - @atlaspack/types@2.15.26
+  - @atlaspack/transformer-raw@2.14.36
+
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`ff8ac84`](https://github.com/atlassian-labs/atlaspack/commit/ff8ac84bd0be15caa7c7af21bf5b6e47b6d36438), [`3da8e69`](https://github.com/atlassian-labs/atlaspack/commit/3da8e690605472d81c531366b65fa9844f8ca337), [`c372f3f`](https://github.com/atlassian-labs/atlaspack/commit/c372f3fd6fce8200d5cf47f41bc7895c6cbb5558), [`27fd73f`](https://github.com/atlassian-labs/atlaspack/commit/27fd73fa07d358653f1be6f2d15d72dc5d8fc2fd), [`3091fa3`](https://github.com/atlassian-labs/atlaspack/commit/3091fa36425ecffe3938e55247a035b08cda37a2)]:
+  - @atlaspack/core@2.26.0
+  - @atlaspack/config-default@7.0.0
+  - @atlaspack/types@2.15.25
+  - @atlaspack/packager-webextension@2.14.35
+  - @atlaspack/runtime-webextension@2.14.35
+  - @atlaspack/transformer-webextension@2.14.35
+  - @atlaspack/transformer-raw@2.14.35
+
 ## 6.0.1
 
 ### Patch Changes